Move IFinders to abstractions
