+++ src/main/java/com/thebois/models/beings/Pawn.java
@@ -4,7 +4,7 @@ import java.util.Collection;
 import java.util.Optional;
 import java.util.Random;
 
-import com.thebois.models.IStructureFinder;
+import com.thebois.abstractions.IStructureFinder;
 import com.thebois.models.Position;
 import com.thebois.models.beings.pathfinding.IPathFinder;
 import com.thebois.models.inventory.items.ItemFactory;
